+++ NEWS.md
@@ -1,3 +1,14 @@
+
+# emphatic 0.1.5  2023-08-30
+
+* Added `hl_str_diff()` to highlight differences between two strings.
+* Added functions for highlighting the results of a `grep()`
+    * `hl_grep()`
+    * `hl_str()`
+    * `hl_character()`
+    * `hl_print()`
+    * `hl_deparse()`
+
 # emphatic 0.1.4  2020-11-13
 
 * Fix for rendering of logical atomic vectors